Only display notification when in a lower version than suggested
