+++ src/browser.js
@@ -8,8 +8,7 @@
 // First step: Always be production, unless told otherwise.
 if (process.env.NODE_ENV === undefined) process.env.NODE_ENV = 'production';
 
-const {ipcRenderer, shell} = require('electron');
-const electron = require('electron');
+const {ipcRenderer} = require('electron');
 const {
   NAVIGATION_REQUESTED,
   RELOAD_REQUESTED,
@@ -170,13 +169,3 @@ function handleLoadAbort(event) {
 function handleLoadRedirect(event) {
   resetExitedState();
 }
-
-const notificationText = {
-  title: 'New Version of the Code.org Maker App Available',
-  body: 'Click this notification to download the latest version of the Code.org Maker App',
-};
-let notification = new electron.remote.Notification(notificationText);
-notification.on('click', (event, arg) => {
-  shell.openExternal('https://studio.code.org/maker/setup');
-});
-notification.show();
